- cleanup. fixed hb_thread_to_integer() on darwin to cast pointer cleanly for both 32-bit and 64-bit.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@2620 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/ports.c
@@ -273,7 +273,7 @@ static uint64_t hb_thread_to_integer( const hb_thread_t* t )
     #if defined( _WIN32 ) || defined( __MINGW32__ )
         return (uint64_t)(ptrdiff_t)t->thread.p;
     #elif defined( SYS_DARWIN )
-        return (uint64_t)(ptrdiff_t)t->thread;
+        return (unsigned long)t->thread;
     #else
         return (uint64_t)t->thread;
     #endif
